(refactor): whitelist and blacklist as transforms
- so that these can work as examples and so that their order is very
  explicit as they're treated like any other transform internally

- create new transforms/ directory that will hold all built-in
  transforms moving forward
  - and export them in transforms/index.ts
