+++ src/index.ts
@@ -1,6 +1,8 @@
 import { onSnapshot, applySnapshot, IStateTreeNode } from 'mobx-state-tree'
 
 import AsyncLocalStorage from './asyncLocalStorage'
+import { ITransform, whitelistKeys, blacklistKeys } from './transforms/index'
+import { StrToAnyMap } from './utils'
 
 export interface IArgs {
   (name: string, store: IStateTreeNode, options?: IOptions): Promise<void>
@@ -12,15 +14,7 @@ export interface IOptions {
   readonly blacklist?: Array<string>,
   readonly transforms?: Array<ITransform>
 }
-type StrToAnyMap = {[key: string]: any}
-
-export interface ITransform {
-  readonly toStorage?: ITransformArgs,
-  readonly fromStorage?: ITransformArgs
-}
-export interface ITransformArgs {
-  (snapshot: StrToAnyMap): StrToAnyMap
-}
+export { ITransform, ITransformArgs } from './transforms/index'
 
 export const persist: IArgs = (name, store, options = {}) => {
   let {storage, jsonify = true, whitelist, blacklist, transforms = []} = options
@@ -39,20 +33,16 @@ export const persist: IArgs = (name, store, options = {}) => {
       'engine via the `storage:` option.')
   }
 
-  const whitelistDict = arrToDict(whitelist)
-  const blacklistDict = arrToDict(blacklist)
+  // whitelist, blacklist, then any custom transforms
+  transforms = [
+    ...(whitelist ? [whitelistKeys(whitelist)] : []),
+    ...(blacklist ? [blacklistKeys(blacklist)] : []),
+    ...transforms
+  ]
 
   onSnapshot(store, (_snapshot: StrToAnyMap) => {
     // need to shallow clone as otherwise properties are non-configurable (https://github.com/agilgur5/mst-persist/pull/21#discussion_r348105595)
     const snapshot = { ..._snapshot }
-    Object.keys(snapshot).forEach((key) => {
-      if (whitelist && !whitelistDict[key]) {
-        delete snapshot[key]
-      }
-      if (blacklist && blacklistDict[key]) {
-        delete snapshot[key]
-      }
-    })
 
     transforms.forEach((transform) => {
       if (transform.toStorage) { transform.toStorage(snapshot) }
@@ -75,16 +65,6 @@ export const persist: IArgs = (name, store, options = {}) => {
 
       applySnapshot(store, snapshot)
     })
-}
-
-type StrToBoolMap = {[key: string]: boolean}
-
-function arrToDict (arr?: Array<string>): StrToBoolMap {
-  if (!arr) { return {} }
-  return arr.reduce((dict: StrToBoolMap, elem) => {
-    dict[elem] = true
-    return dict
-  }, {})
 }
 
 function isString (value: any): value is string {
