+++ src/transforms/blacklist.ts
@@ -1,0 +1,14 @@
+import { ITransform, arrToDict } from './utils'
+
+export function blacklistKeys (blacklist?: Array<string>): ITransform {
+  const blacklistDict = arrToDict(blacklist)
+
+  return {toStorage: function blacklistTransform (snapshot) {
+    Object.keys(snapshot).forEach((key) => {
+      if (blacklist && blacklistDict[key]) {
+        delete snapshot[key]
+      }
+    })
+    return snapshot
+  }}
+}
