+++ src/utils.ts
@@ -1,0 +1,1 @@
+export type StrToAnyMap = {[key: string]: any}
